src/scripts/requirements.txt
mkdocs
mkdocs-material
mkdocs-mermaid2-plugin
mkdocs-include-dir-to-nav
mkdocs-awesome-pages-plugin
mkdocs-redirects
pandas
pyyaml
tabulate
requests
lxml
bs4
openpyxl
Pillow